move tableDataSlice to slices
